backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
langchain>=0.3.14
langchain-openai>=0.3.0
langchain-pinecone>=0.2.3
langchain-text-splitters>=0.3.4
pinecone-client>=6.0.0
pypdf2>=3.0.1
python-dotenv>=1.0.1
pydantic>=2.10.5
pydantic-settings>=2.7.1
sqlalchemy>=2.0.0
PyJWT>=2.8.0
psycopg2-binary>=2.9.9
python-docx>=1.1.0
python-pptx>=1.0.2
